Reorganizing sidebar files
