+++ frontend_pages/new_project/ui_new_project_window.py
@@ -17,7 +17,7 @@ from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor,
     QPalette, QPixmap, QRadialGradient, QTransform)
 from PySide6.QtWidgets import (QApplication, QSizePolicy, QWidget)
 
-from widgets.sidebar import Sidebar
+from widgets.sidebar.sidebar import Sidebar
 
 class Ui_Form(object):
     def setupUi(self, Form):
